rename route config
